Changed logical location of exit command.
If an upload script has been added, in the event that an upload fails The upload will be attempted again if there are no updated versions of FireFox available upon the next run of the build.bash file. This means that in the event of the upload failing it will be attempted upon the next run, which is a good idea when this build is schooled on a regular bases.